Allow a foreword comment under the header of a group in a bibliography
An empty group without a title but with a foreword can serve as
a preface to the whole bibliography section when put in front of all
other groups, or as a transition comment when put in between of two
groups in the middle.

The text is interpreted as plain Markdown (no shortcodes).