Allow interactive filtering of bibliography entries
Add a straightforward text-based filtering of bibliography entries.
On every change of the contents of the filter text field, split the
query into whitespace-separated words and hide all the entries that
do not have occurrences of all of the words as substrings somewhere
within them. The list of fields that are scanned for matches in each
entry is found in the 'bib.entryFieldsSelector' array (see
static/js/bibliography.js). Differences in case are ignored.

Although the performance seems to be satisfactory for the foreseeable
future, it might perhaps be justified to build a Trie-based index or a
DAFSA-based matcher on page load.