Enhance bibliography filtering by separate category selection
Although exact type is known for each entry, to simplify the choice
for the user, group all entry types into four categories: “conferences”,
“journals”, “books”, and “other”. See the exact groupings in the
definition of the ‘entryKindSelectionMap’ in bibliography.js.

In order not to parse the ‘class’ attribute of the entry tags in the
filtering code, an extra ‘data-bib-kind’ attribute with the BibTeX
type is added to the list item tags. (Attribute whose names are
prefixed with ‘data-’ are reserved in HTML for custom data private
to the page or application.)