Add text to show for no matches in bibliography filtering
It may be confusing for the user to see nothing under the bibliography
header when there is no entry matching the entered filter text (especially
if the user has forgotten that they typed in a query in the first place).
To avoid that, show a dummy bibliography group with the 'No match'
title in case all entries end up being made invisible.