Add links to personal pages from bibliography entries
Whenever a list of author or editor names is generated to make
a bibliography entry (and not a citation), take the full name of each
author or editor, and look through all the personal pages in search of
one with that same name (sans any differences in Unicode whitespace
characters, in case a person would like to have, say, the non-breaking
space instead of a regular one in their name on the personal page).

For all links to be found and generated perfectly, it is required that
(1) the name field in the metadata of a personal page lists the full
name of the person that they also use in the BibTeX entries for their
publications;
(2) there is no two people with the same full name in the lab;
(3) there is no external collaborators in the BibTeX entries that bear
the same full name as a person in the lab.

Nothing breaks if any of these three assumptions is violated, but some
author or editor links might be missing (if (1) or (2) does not hold),
or might be leading to the wrong person (if (2) or (3) does not hold).