Support links to external collaborators from bibliography entries
Similarly to linking the personal pages of lab members, also create
links in bibliographies for frequent external collaborators by matching
the full names from the author or editor lists in BibTeX entries with
the items in the 'collaborators' array in the global bibliography settings
(see 'bibliography' parameters under the website configuration in the
'config.yaml' file). Each entry in the 'collaborators' array must have
the 'name' and 'url' fields, containing the full name of a collaborator
and the fully-qualified URL to their web page, respectively.