scripts/bibtex2yaml: Add an option to reformat the input BibTeX file
The fields within a BibTeX entry are placed in the canonical order:
standard fields first, in the order given by the BIBTEX_STANDARD_TAGS
tuple, followed by non-standard fields in alphabetical order.

The entries are sorted in reverse chronological order.

None of the fields have their content modified for now;
this may change in the future.