scripts/bibtex2yaml: Expand @string{} variables in BibTeX source snippets
In order to preserve the original content of the input BibTeX file
when reformatting it, we keep ‘interpolate_strings=False’ in the
‘BibTexParser’ constructor. So, if the input file has @string{}
variables, say, for journal names, the definitions of the entries
with the ‘journal’ field using those variables stay unaffected.

However, we also generate separate BibTeX snippets for each entry,
which are intended for an outside user who, most likely, does not
structure their BibTeX files the same way, and might not even use
any @string{} variables. Hence, to make each snippet self-contained,
we expand variables into their values when generating snippets.

We make an exception for the ‘month’ field, though, as it is very
frequent (and, in fact, recommended) for the field to use the month
variables. Since those variables are defined by default by all BibTeX
processors, there is no risk of making the entry not self-contained
because of that. But beware: we do not interpolate *any* variables
in the ‘month’ field, so if, in an unlikely case, some other variable
is used within it, that variable will not be interpolated, too.