scripts/bibtex2yaml: Correct name splitting to respect brace-protected portions
Proper interpretation of LaTeX subexpressions enclosed in braces
is required for correct parsing of names in the ‘author’ and ‘editor’
fields. For instance:
• if an author is an organization, its title should be braced in order
not to be incorrectly broken into multiple parts as a person’s name;
• if a name has the word ‘and’ in it (e.g., in case of the previous
point), the containing part of the name must be braced;
• if a part of a name (e.g., the first or last name) consists
of multiple words, they should be braced to be kept together;
• if a name has special characters in it, a braced subexpression
within a word might be used to generate those characters; etc.

The heavy-lifting of parsing a single name is already done by the
splitname() function in the ‘bibtexparser’ package, but the overall
correctness depends on the name list being split into separate name
expressions correctly.

Following the changes in parsing of the ‘title’ field, we can now extend
the use of the ‘pylatexenc’ package for parsing the content of the
‘author’ and ‘editor’ fields. The original attempt of splitting literally
by the space-surrounded ‘and’ word may interfere with braced
subexpression containing ‘and’. Parsing of the whole expression
allows to correctly preserve the name fragments protected with
braces, and pass them on to splitname() for proper interpretation.