scripts/bibtex2yaml: Support the 'others' keyword in BibTeX name lists
BibTeX allows for the ‘others’ keyword to be used within the ‘author’
or ‘editor’ fields as the literal “et al.” placeholder for omitted names
(as opposed to the one generated automatically for cutting long name
lists short).

When the keyword is found in a name list, it is correctly passed on to
the dictionary in the corresponding name array inside the output YAML
bibliography data record. The dictionary of such pseudo-person entry
has only its ‘last’ field set—to ‘others’. Regardless of where it appears
inside the input name list, it is always put at end in the resulting
name array.