Support the 'others' keyword in name lists in bibliography data
The keyword has the same “et al.” meaning as in BibTeX, i.e., stands
for a number of omitted names in the ‘author’ and ‘editor’ fields.
Hence, it is converted into the ‘et al.’ mark when bibliographies
are generated—but only when it is the last “name” in the list.

The keyword should never be used in bibliography entries of our own
publications, but is supported just in case it appears in an entry of
an external publication. But even in that case, it should not be used
unless the missing authors are indeed unknown (and not omitted just
out of laziness).

The bib/getfullname.html partial is now used to check whether a name
profile has nothing but the ‘last’ field set. For this reason, it is refactored
so that such profile is rendered exactly to the content of that ‘last’ field.
(Previously, it always included the content of the ‘first’ field even when
it was unset, erroneously making the returned full name start with a
superfluous ‘<nil>’ and a space in that case.)