Update config for the switch to Goldmark Markdown processor
Starting with Hugo 0.60, the default Markdown processor is changed
from Blackfriday to Goldmark (due to its superior performance).
It makes sense to follow the new default choice.

The two processors have minor differences in rendering. The only
noticeable regression of the switch on our pages is due to Goldmark
by default omitting embedded HTML, which we use in the source of
our landing page (see content/_index.md). This commit reverses the
corresponding setting in the configuration.