scripts/bibtex2yaml: Keep empty lines between paragraphs in detexify()
There is a slight potential disadvantage to this change. We do not
separate the LaTeX-formatted BibTeX fields into those that are
essentially one-line only (e.g., ‘title’) and those that may contain
a piece of text, generally, of multiple paragraphs (e.g., ‘abstract’).
This means that if a blank line is introduced into the value of
a naturally one-line field, it will be preserved. Given that mistakes
like that should be easily catchable, though, maintaining which field
is which in the script does not seem to be worth it at the moment.