scripts/fetchsite: Reset to origin before pulling the changes
In a rare case of the tracked branch being force-pushed, a pull by
itself will not be enough to adopt the potentially conflicting new
commits. Further, since the local and remote hashes will remain
different after the script terminates, if it is run repeatedly (e.g.,
on schedule as a cron job), excessive fetching attempts will be
induced on its future invocations. To cover for such cases, make
sure to reset to the (new) origin’s state prior to the pull (but after
the fetch).