Eliminate uses of .Scratch for plain variables in bibliography lookup
In early versions of Hugo, the only way to have a state maintained
between different iterations of a ‘range’ loop was to use the per-page
‘.Scratch’ dictionary. We had to resort to that approach to implement
bibliography item lookup for citing publications and pulling up their
reference information.

Starting with Go 1.11, variable overwriting is supported in templates.
Since version 0.48, Hugo is built with Go 1.11 or newer, and therefore
inherits this feature for the layout templates. With overwriting allowed,
all our uses of ‘.Scratch’ can be simplified with plain local variables.