Remove the '.html' filename extension from pure-function partials
Partials whose only purpose is to compute and return a value are
significantly different in their nature from regular ones whose purpose
is to produce a piece of content for substitution at the call site.

We already have a naming convention to differentiate the former by
prefixing their names with ‘get’. Further, it does not make sense for
their names to have any content-based suffix (as their outcome is
a value, such as a dictionary or a string, and not a fragment of any
formatted content).