people: Remove the 'Past Visitors' header for the visitors list
Now that we have some visitors with personal pages, there is already
a header saying ‘Visitors’ (leading the profile cards of such visitors).
Since the Visitors section is always the last one in the automatically
populated part of the People page, given that it exists, the manually
added content from the ‘_index.md’ file is attached right after it.
The visual proximity naturally prompts for reusing the above Visitors
section header for both pieces of content.

This might (perhaps even should) change in the future, if we start
using the content from the ‘_index.md’ file for information other than
the past visitors list.