Minify CSS and JS resources before linking from the head
To reduce the number of requests required to fetch the stylesheets,
join them into a single one. To save a bit of traffic, minify the result.
Minify the (currently) only JS file, as well. The resulting bundles
are fingerprinted to make sure the browser does not cling to an old
version of a bundle from cache when it is changed.

In order to implement the said transformations via Hugo resources.*
methods, we have to move the corresponding subtrees of files from
the static/ to assets/ directory.