Emphasize an active main menu item with a heavier underline
A menu item is rendered active on the current page if it belongs
to the section that item is linked to, as evidenced by its URL being
a prefix of the current page URL.

For the check to be correct, the former URL has to have a slash at the
end, if it does not have one. (Otherwise, it will be incorrect if the
website has a section whose name is a prefix of another section’s name.)
We make sure to add that slash before the check, if necessary.