Ensure Unix line endings in checked-in text files via '.gitattributes'
Just in case somebody will be making edits to the site‘s content
on Windows, the directive

	*	text=auto

in the '.gitattributes' file converts all checked-in text files in the
repository to Unix line endings from the default line endings used
by the user’s operating system. This does not require the user to
change anything in their local Git configuration.

Git does not perform this automatic line-endings conversion on files
that it detects to be binary. To make sure that files of the types
commonly used in our repository are always detected as text or binary
correctly, we also add the corresponding explicit directives (based on
file extensions) to the '.gitattributes' file.