scripts/fetchsite: Use a shallow repository of single-commit depth
In order to minimize the storage size and network load, try to keep the
repository in the shallow state, where only the information required to
reconstruct the latest commit is requested and stored.

Therefore, pass the --depth=1 option to Git when doing a fetch, pull,
or clone. Since --depth assumes --single-branch during a clone, the
latter option is removed from the clone command as superfluous.