scripts/fetchsite: Insist on a base URL only if the -u option is given
If Hugo is not given the --baseURL option, it uses the base URL set
in the configuration file of the website. So, rather than insisting on
a localhost base URL when the -u option is not given to the script,
just do not pass any --baseURL options to Hugo in that case.