Switch to a mapping-based representation of bibliography data
Previously, we used a sequence-based format, where the records were
output as a list (under the single top-level 'references' key). To simplify
the subsequent lookup of the resulting data, switch to a mapping-based
format, where the records are output as a dictionary (still under that
'references' key) whose keys are the records’ BibTeX ids and values
are the same records as before, but with the 'id' fields elided.

In order to complete the transition, in this commit we:
1. Rewrite 'scripts/bibtex2yaml' to adopt the new mapping-based format.
2. Regenerate existing bibliography data files from the BibTeX sources
by rerunning the 'mkbibdata' script.
3. Adjust the bibliography entry lookup in the 'bib/getentrydata' partial
to switch from using 'range' loops in favor of simple 'index' calls.

This change of the format allows for more performant bibliography
entry lookup that we can benefit from during the generation of all
citations and publications lists across the website.