Switch to the itemized name format in the metadata of personal pages
In preparation to the downstream changes for eliminating potential
ambiguity in bibliography name lookup, change the full name format
in the personal pages’ metadata from a single string to a dictionary
with separate keys for the first, middle, last names, and the lineage
suffix, i.e., use exactly the same format as in bibliography data files.

The motivation for this change is to make component-wise comparison
of names possible (between personal pages and also with bibliography
entries), mainly for bibliography name lookup and sorting.