Use last names first for sorting profiles by name on the People page
When sorting people’s profiles by name, take advantage of the recent
change to the itemized name format in the personal pages’ metadata,
and compare first by last name, then by lineage suffix, then by first
name, and finally by middle name. (It was intended to be this way from
the beginning, but could not be easily done when only a single-string
full name was available for each person.)

To ensure correct ordering for multi-key sorting of pages, we rely on
the ‘ByParam’ method instead of the ‘sort’ function, as the former,
unlike the latter, uses a stable sorting algorithm. Thus, ‘ByParam’
is called multiple times, each time with the next most significant
sorting key.