Support the 'bibname' metadata field for bibliography name linking
It is now possible to specify an optional ‘bibname’ field in the personal
pages’ metadata, to be used for linking a personal page to an author
or editor name in bibliography entries. This new field has the same
itemized format as the ‘name’ field, with the latter being used, as
before, in case the former is missing.

The ability to separate the two kinds of names is added to make sure
that the bibliography name linking does not break in case a person
prefers for their website display name not to match the full name they
use for publishing and bibliography referencing purposes (e.g., when
the person uses their middle name only in the form of an initial normally,
but in its full form in publications’ data).