Lookup personal pages by unique name IDs instead of full names
When rendered into a human-readable string, a full name may not
be sufficient to disambiguate between two people with multi-word
first or last names. For instance, the names

	name:
	  first: John
	  middle: Cohen
	  last: Smith

and

	name:
	  first: John
	  last: Cohen Smith

result in the same full name.

To eliminate this potential ambiguity, generate unique name ID strings
(not for human consumption) built from the itemized name dictionaries,
which are now consistently used both in bibliography data and personal
pages’ metadata to specify names. Possible fields in those dictionaries
are defined in the ‘namedparts’ property of the ‘bibliography’ section
in the website configuration file. Each part is defined by the key to be
indexed from the name dictionary and the priority it has for sorting
(the parts that come first in their priority values are compared first).

The name ID is constructed as the concatenation of the name parts,
in the order of priority, separated with null bytes, which cannot be
a part of any valid UTF-8 code point and can, therefore, be safely
used as a separator. If a certain name part is absent, an empty string
is concatenated at the corresponding position in the name ID. This
way, it can be guaranteed that two names are component-wise identical
if and only if they have identical name IDs.

For the above two examples, the corresponding name IDs are

	"Smith\x00\x00John\x00Cohen"

and

	"Cohen Smith\x00\x00John\x00"

respectively (in the Go string literal syntax).

Furthermore, thanks to the priority ordering of the name parts within,
name IDs can be used as keys for proper sorting by full name.