Cache the name ID-to-collaborator mapping in the people index object
Simplify the bibliography name lookup for collaborators in the same
way it is done for personal pages. Namely, build a dictionary for
collaborators prior to rendering a bibliography section in the
bib/getpeopleindex partial (called from the bib/section.html one).

The people index, therefore, is extended to have two fields, both
dictionaries: ‘internal’ with references to personal pages as values,
and ‘external’ for collaborator description objects from the website
configuration as values. Both dictionaries use name IDs as keys.