Disable RSS feeds for all page kinds for now
The automatically generated RSS feeds (which were enabled for the home
page and all section pages by default prior to this change) are not populated
to look like anything we might want to see in a news feed for the website.
Hence, disable RSS output until we have a good idea of what should be
put into the feed(s).