scripts/gethugo: Remove support for the Windows OS suffix
Windows builds of Hugo are released as .zip archives (unlike .tar.gz
archives for Unix-like platforms), which require different handling of
the downloaded file. The necessary complication of the script is not
considered justified for now.

Note that the script should still work correctly when it is run under
the Windows Subsystem for Linux (WSL) on Windows 10 or later,
because the Bash shell in WSL sets $OSTYPE to ‘linux-gnu’.