Support source code links in bibliography entries
Allow authors to publicize a link to the source code directly related
to a publication next to its bibliography entries on the site. For
that, a non-standard ‘code’ field is supported in BibTeX records.
If a link is provided in the value of that field, it is added in the
row of extras following the corresponding bibliography entry.