Recognize links to arXiv preprint releases in bibliography entries
In addition to arXiv, we also recognize ‘eprinttype’s of another three
digital libraries: HDL, JSTOR, and PubMed—the same ones that are
supported by the BibLaTeX extended styles (see the documentation
of the ‘biblatex-ext’ LaTeX package).

The URL templates for each of the four ‘eprinttype’s are looked up (in
lowercase) in the ‘eprint’ dictionary under the ‘urltemplates’ section
of the ‘bibliography’ parameters in the website configuration.

For consistency, the URL template for DOI links is also moved into the
‘doi’ entry under that same ‘urltemplates’ section of the ‘bibliography’
parameters.