Change the mark for starred bibliography entry keys to the ‘*’ suffix
Previously, we required bibliography keys of starred entries to be
prefixed with ‘+’, instead. This was due to the ‘+’ character being
one of the few that do not have a special meaning at the start of an
unquoted string value in YAML—unlike the more natural ‘*’ character
that does—and we value the convenience of not being required to quote
BibTeX keys in the YAML-formatted page metadata.

However, having the ‘*’ character at the end of a string (unlike the start)
does not require quoting it. Hence the switch from the ‘+’ prefix to the
‘*’ suffix as the more mnemonic syntax. (We believe that it will be
easier for users to remember that starred entries are marked with
a star rather than any other character.)