Extract paths to bibliography entry attachments into the config
This new parametrization allows for better configurability of the
paths to be inspected under the static/ directory in search of the
files to be attached to each bibliography entry, based on its id.

For example, in this commit we add support for supplement paper
attachments and make it possible for the slide deck attachments
to be Microsoft PowerPoint, Apple Keynote, or OpenDocument
Presentation files.