Extract display settings defined by people roles into the config
Currently, there are three ways in which a role changes the display of
information from personal pages assigned to it in the directory on the
People page. Namely, for each personal page, it determines:
• at which position relative to other personal cards does this page’s
card show up within its role directory;
• whether the (first) website given in the page’s metadata is featured
in this page’s card in the directory;
• whether academic year hints are provided for the half-open year
ranges on this page and its card in the directory.

Previously, these display properties were determined by inspecting
the ‘role’ attribute right next to the template code affected by them.
Now, we delegate these decisions to the three new fields in the role
configuration dictionaries: ‘sortby’, ‘featurewebsite’, and
‘academicyearhint’, respectively.

The latter two fields are self-evident Boolean flags.

The ‘sortby’ field provides the list of metadata attributes to serve
as sorting keys for ordering the personal pages’ cards on the People
page. The attributes are listed in the order of decreasing significance,
with each successive attribute breaking the ties between pages with
identical values of the preceding ones. The ‘name’ attribute induces
correct multi-component comparison of full names.

Each attribute name given in the ‘sortby’ list may be prefixed with
the ‘↑’ or ‘↓’ character to indicate that the ascending or descending
order of sorting is requested for this attribute, respectively (with
the former being the default if neither of the prefixes is present).