Unify the format of the copyright and contact parameters in the config
Switching from the email-only format of the contact link allows for
more flexibility in terms of the means of contact. The generic URL
field makes it possible to link to a mailing list, a messenger channel,
a social network page, etc. To achieve the same effect as before,
simply provide a pertinent ‘mailto:’ link as the URL.